Use a shared volume to store app gems + add aliases
Configure a shared volume on the Docker host.
Mount it inside the Docker image.
Set the proper right with chown.
Add vscode user bash_aliases inside the volume with a symlink instead of copying the file to the /home/vscode
This allow to edit it easily from host and to load and link this repository .devcontainer folder via a symlink in futur rails projects to share the configuration

Ruby, rails, rubocop will be installed on the container inside rbenv by default
Then later, each project Gemfile gems will be installed inside /bundle
This allow to rebuild the Dev container at any time with no need to reinstall the whole project setup (bundle install...)
You might still need to reconfigure a database